+++ src/clonesvg.js
@@ -1,6 +1,7 @@
 /* Some utilities for cloning SVGs with inline styles */
 import computedStyles from 'computed-styles';
-import {isUndefined} from './utils';
+import {isUndefined, isObject, clone} from './utils';
+import {inheritableAttrs} from './collection';
 
 // Removes attributes that are not valid for SVGs
 function cleanAttrs (el, attrs, styles) {  // attrs === false - remove all, attrs === true - allow all
@@ -18,19 +19,38 @@ function cleanAttrs (el, attrs, styles) {  // attrs === false - remove all, attr
     });
 }
 
-// Clones an SVGElement, copyies approprate atttributes and styles.
-export function cloneSvg (src, attrs, styles) {
-  var clonedSvg = src.cloneNode(true);
-  var srcChildren = src.querySelectorAll('*');
-
-  computedStyles(src, clonedSvg.style, styles);
-  cleanAttrs(clonedSvg, attrs, styles);
-
-  Array.prototype.slice.call(clonedSvg.querySelectorAll('*'))
-    .forEach(function (target, index) {
-      computedStyles(srcChildren[index], target.style, styles);
-      cleanAttrs(target, attrs, styles);
+function cleanStyle (tgt, parentStyles) {
+  if (tgt.style) {
+    inheritableAttrs.forEach(function (key) {
+      if (tgt.style[key] === parentStyles[key]) {
+        tgt.style.removeProperty(key);
+      }
     });
+  }
+}
+
+function walker (attrs, defaultStyles) {
+  return function walk (src, tgt) {
+    if (!tgt.style) return;
+
+    computedStyles(src, tgt.style, defaultStyles);
+
+    const children = src.childNodes;
+    for (var i = 0; i < children.length; i++) {
+      walk(children[i], tgt.childNodes[i]);
+      cleanStyle(tgt.childNodes[i], tgt.style);
+    }
+
+    if (tgt.attributes) {
+      cleanAttrs(tgt, attrs, defaultStyles);
+    }
+  };
+}
+
+// Clones an SVGElement, copies approprate atttributes and styles.
+export function cloneSvg (src, attrs, styles) {
+  const clonedSvg = src.cloneNode(true);
+  walker(attrs, styles)(src, clonedSvg);
 
   return clonedSvg;
 }