+++ CHANGELOG.md
@@ -2,7 +2,7 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+* Clean superfluous inheritable styles
 
 --------------------
 
@@ -18,4 +18,3 @@ _(none)_
 ## 0.3.0 (2015-10-20)
 * Now using copy-styles module
 * No longer removes styles from elements that match parent styles (not all styles are inheritable)
-